Whatever i did, it seems to be working
For some reason, the public sources page now uses the poster terminology instead of the image terminology, which is the opposite of which I wanted to do it'll still work because it's standalone, but I don't know it's weird. I don't know how to do this and probably I might've changed that at some point and I just didn't change it back. This is a pain.